PR target/23473 	* arm.md (arm_load_pic_register): Change argument to the mask of 	saved registers.  Call thumb_find_work_register if we need a 	scratch register on Thumb. 	(arm_expand_prologue): Pass empty register set to 	arm_load_pic_register. 	(thumb_expand_prologue): Pass live_regs_mask directly to 	arm_load_pic_register. 	* arm-protos.h (arm_load_pic_register): Update prototype.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@103277 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/config/arm/arm-protos.h
@@ -27,7 +27,7 @@
 extern void arm_override_options (void);
 extern int use_return_insn (int, rtx);
 extern int arm_regno_class (int);
-extern void arm_load_pic_register (unsigned int);
+extern void arm_load_pic_register (unsigned long);
 extern int arm_volatile_func (void);
 extern const char *arm_output_epilogue (rtx);
 extern void arm_expand_prologue (void);
